Team Goal Timer is now handled server side
This was always the plan. I was originally handling the update after each EndClientTurn (like the other timers), and for testing, this was okay. But once you have multiple teams with more than one or two members, I knew it would be horribly inefficient because it doesn't need to updated by every client in the alliance.